DPE-836: Refactored code as per feedback, honoring kubeconfig and context passed in all commands.
